requirements.txt
# Python3.7
# CUDA 10.0
# cuDNN 7.6.5 for CUDA 10.0
# 
# TensorFlow 1.14
# Deepgaze

absl-py==0.9.0
astor==0.8.1
click==7.1.2
coloredlogs==14.0
deepgaze==0.1
dlib==19.20.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
humanfriendly==8.2
importlib-metadata==1.7.0
imutils==0.5.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.16.0
opencv-python==4.3.0.36
pandas==1.0.5
Pillow==7.2.0
protobuf==3.12.2
python-dateutil==2.8.1
pytz==2020.1
scipy==1.5.1
six==1.15.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
ujson==3.0.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
